backend/requirements.txt
# Core FastAPI dependencies
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
pydantic==2.9.2
pydantic-settings==2.6.1

# Document processing - Updated to latest stable version
docling==2.53.0
transformers>=4.44.0,<4.46.0
torch>=2.1.0,<2.7.0
torchvision>=0.16.0,<0.20.0
Pillow>=10.0.0

# Additional ML dependencies for Granite Docling
accelerate>=0.25.0
sentencepiece>=0.1.99
protobuf>=4.21.0

# File handling and utilities
aiofiles>=23.2.0
python-magic>=0.4.27

# HTTP and async
httpx>=0.25.0
aiohttp>=3.9.0

# Logging and monitoring
structlog>=23.2.0
python-json-logger>=2.0.0

# Development and testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
black>=23.11.0
isort>=5.12.0
flake8>=6.1.0
mypy>=1.7.0

# Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Environment management
python-dotenv>=1.0.0

# Additional dependencies
markdown>=3.9
python-magic-bin>=0.4.14  # Windows-compatible version of python-magic

# CORS support (Note: CORS is built into FastAPI, this package may not be needed)
# fastapi-cors==0.0.6
